requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anthropic==0.21.3
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.15.0
azure-search-documents==11.4.0
azure-storage-blob==12.20.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.1
confection==0.1.4
contourpy==1.2.1
cryptography==42.0.5
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docx2txt==0.8
docxcompose==1.4.0
docxtpl==0.16.8
emoji==2.11.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
extra-streamlit-components==0.1.71
Faker==24.9.0
favicon==0.7.0
filelock==3.13.3
filetype==1.2.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.22.0
idna==3.6
importlib-metadata==6.11.0
intel-openmp==2021.4.0
ipykernel==6.29.3
ipython==8.22.2
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
langchain==0.1.11
langchain-anthropic==0.1.4
langchain-community==0.0.27
langchain-core==0.1.30
langchain-google-genai==0.0.11
langchain-text-splitters==0.0.1
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.1.23
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.2.0
mpmath==1.3.0
msal==1.27.0
msal-extensions==1.1.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.13.3
openpyxl==3.1.2
orjson==3.9.15
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
portalocker==2.8.2
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pymdown-extensions==10.7.1
PyMuPDF==1.24.3
PyMuPDFb==1.24.3
pyotp==2.9.0
pypandoc==1.13
pypandoc-binary==1.12
pyparsing==3.1.2
pypdf==4.1.0
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
pytz==2024.1
# pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
rapidfuzz==3.8.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.28
srsly==2.4.8
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
streamlit==1.36.0
streamlit-authenticator==0.2.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-chat==0.1.1
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-feedback==0.1.3
streamlit-flexselect==0.0.6
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-navigation-bar==3.3.0
streamlit-option-menu==0.3.6
streamlit-quill==0.0.3
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.2.3
thinc==8.2.3
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.40.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
unstructured==0.13.2
unstructured-client==0.18.0
urllib3==2.2.1
validators==0.24.0
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.13
weasel==0.3.4
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
